requirements.txt
einops==0.8.1
matplotlib==3.8.0
numpy==1.26.4
pandas==2.3.0
Pillow==11.2.1
PyYAML==6.0.2
scikit_learn==1.2.2
scipy==1.15.3
timm==0.9.12
torch==2.2.0
torchinfo==1.8.0
